!B (Sandbox) (CE-19402) ROAD: The width of a road's spline/point should be changeable as floats, not as ints
Copied from Perforce
 Change: 1909053

+++ Code/Sandbox/EditorQt/Objects/SplineObject.cpp
@@ -1171,8 +1171,8 @@ void CSplineObject::SerializeProperties(Serialization::IArchive& ar, bool bMulti
 
 			if (!isDefault)
 			{
-				int pwidth = GetPointWidth();
-				int oldWidth = pwidth;
+				float pwidth = GetPointWidth();
+				float oldWidth = pwidth;
 
 				ar(pwidth, "pwidth", "Width:");
